requirements.txt
torch
torchvision
opencv-python
pandas
matplotlib
scikit-learn
ultralytics
albumentations
pillow
tqdm
pyautogui
keyboard
numpy